+++ src/main/java/com/tttm/birdfarmshop/Config/SecurityConfigure.java
@@ -1,4 +1,4 @@
-package com.tttm.birdfarmshop.config;
+package com.tttm.birdfarmshop.Config;
 
 import com.tttm.birdfarmshop.Models.ERole;
 import lombok.RequiredArgsConstructor;
@@ -8,11 +8,9 @@ import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfiguration;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.filter.CorsFilter;
 
 @Configuration
 @EnableWebSecurity
