+++ src/main/java/com/tttm/birdfarmshop/Controller/AuthController.java
@@ -1,24 +1,31 @@
-package com.tttm.birdfarmshop.controller;
+package com.tttm.birdfarmshop.Controller;
 
 
 import com.tttm.birdfarmshop.Models.User;
 import com.tttm.birdfarmshop.Service.AuthenticationService;
 import com.tttm.birdfarmshop.Utils.AuthenticationRequest;
 import com.tttm.birdfarmshop.Utils.AuthenticationResponse;
-import com.tttm.birdfarmshop.exception.CustomException;
+import com.tttm.birdfarmshop.Exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/admin")
-public class AdminController {
+@RequestMapping("/auth")
+public class AuthController {
     private final AuthenticationService authenticationService;
+    @PostMapping("/register")
+    public ResponseEntity<AuthenticationResponse> register(@RequestBody User user) throws Exception {
+        try {
+            return new ResponseEntity<>(authenticationService.register(user), HttpStatus.OK);
+        }
+        catch (Exception ex)
+        {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+    }
 
     @PostMapping("/login")
     public ResponseEntity<AuthenticationResponse> login(@RequestBody AuthenticationRequest user) throws CustomException {
