+++ src/main/java/com/tttm/birdfarmshop/Service/AuthenticationService.java
@@ -4,8 +4,7 @@ package com.tttm.birdfarmshop.Service;
 import com.tttm.birdfarmshop.Models.User;
 import com.tttm.birdfarmshop.Utils.AuthenticationRequest;
 import com.tttm.birdfarmshop.Utils.AuthenticationResponse;
-import com.tttm.birdfarmshop.exception.CustomException;
-import org.springframework.http.ResponseEntity;
+import com.tttm.birdfarmshop.Exception.CustomException;
 
 public interface AuthenticationService {
     AuthenticationResponse register(User dto) throws CustomException;
